Create Week 0: Syllabus
Course Resources

In general, the class will be organized around the material provided in the required text, but this will be supplemented by additional material, including, but not limited to additional readings and video lectures (as replacements for the in-class lectures). It is important to note that some material in the textbook will not be discussed in my video lectures. While I will focus on the material I believe is most important, do not assume that other assigned reading material will not be covered on exams. 

Required Materials - EBOOK:
Title 	Edition 	Author 	Publisher 	ISBN 
Personality  Psychology 	6th 	Larsen & Buss	McGraw-Hill Education 	978-1-259-87049-1

For this course you will be required to purchase McGraw-Hill Education with Connect® access for Personality Psychology: Domains of Knowledge about Human Nature, 6th Edition.
You are not required to have a print text
Connect access can be purchased during registration for the program, or a connect access card can be purchased from our campus bookstore. The ebook version of your text is included within Connect. To register for Connect, please click on your first Connect assignment in Blackboard and you will be directed to the registration page, or access the connect section directly here: https://connect.mheducation.com/class/a-sutu-3325-spring-2021 Use your UH email to sign in!! 
		
Optional Textbook 

Note: I will use this book for class slides and some readings. Classes are already expensive, and I am not a fan of financial barriers to education. As such, you do not need to purchase this book. I will share with you the readings from the book! 
Funder, D. C. (2015). The Personality Puzzle (7th edition). W. W.  Norton & Company.
If you want to peruse this book, I will gladly lend you a copy of it. Let me know in advance, so we can arrange a socially-distanced pick-up.